Fix error handling bug when signal api emits error
When receiving messages and the Signal API returns an error, we were not
correctly listening to the 'error' event handler, this resulted in us
hanging and waiting forever.

Now we catch the errors via the event handler, and propagate that back
up to the controller so the message can be seen by the user.